finished the backend, will work on the backend ltr

+++ Backend/app/crud.py
@@ -1,4 +1,5 @@
 from sqlalchemy.orm import Session
+from fastapi import HTTPException
 from .  import models, schemas
 
 def get_tasks(db: Session, skip: int = 0, limit: int = 10):
@@ -11,9 +12,27 @@ def create_task(db: Session, task: schemas.TaskCreate):
     db.refresh(db_task)
     return db_task
 
-def delete_task(db: Session, task: schemas.TaskDelete):
-    db_task = models.Task(**task.model_dump())
+def delete_task(db: Session, task_id: int):
+    db_task = db.query(models.Task).filter(models.Task.id == task_id).first()
+    if db_task is None:
+        raise
+    HTTPException(status_code=404, detail="Task not found")
     db.delete(db_task)
+    db.commit()
+    return db_task
+
+def update_task(db: Session, task_id: int, task: schemas.TaskUpdate):
+    db_task = db.query(models.Task).filter(models.Task.id == task_id).first()
+    if db_task is None:
+        raise HTTPException(status_code=404, detail="Task not found")
+
+    for key, value in task.dict(exclude_unset=True).items():
+        setattr(db_task, key, value)
+
     db.commit()
     db.refresh(db_task)
     return db_task
+  
+    
+    
+
